+++ lib/ooxml_parser/docx_parser/docx_data/document_structure/docx_paragraph.rb
@@ -315,7 +315,7 @@ module OoxmlParser
           paragraph_style.style = StyleParametres.parse(style)
         end
         style.xpath('w:rPr').each do |r_pr|
-          DocxParagraphRun.parse(r_pr, character_style, @default_character)
+          DocxParagraphRun.parse(r_pr, character_style, DocumentStructure.default_run_style)
         end
         break
       end