+++ lib/ooxml_parser/docx_parser/docx_data/document_structure/header_footer.rb
@@ -1,37 +1,56 @@
 module OoxmlParser
+  # Class Header Footer classes
   class HeaderFooter < OOXMLDocumentObject
-    attr_accessor :id, :description, :type
+    attr_accessor :id, :elements, :type
+    attr_accessor :path_suffix
 
     def initialize(type = :header)
       @type = type
+      @elements = []
     end
 
-    def parse(type, id, default_paragraph, default_character)
-      if type == :header
-        doc = Nokogiri::XML(File.open(OOXMLDocumentObject.path_to_folder + 'word/footnotes.xml'))
-        doc.search('//w:footnote').each do |footnote|
-          next unless footnote.attribute('id').value == id
-          description = []
-          paragraph_number = 0
-          footnote.xpath('w:p').each do |paragraph|
-            description << DocxParagraph.parse(paragraph, paragraph_number, default_paragraph, default_character)
-            paragraph_number += 1
-          end
-          return description
-        end
-      elsif type == :header
-        doc = Nokogiri::XML(File.open(OOXMLDocumentObject.path_to_folder + 'word/endnotes.xml'))
-        doc.search('//w:endnote').each do |footnote|
-          next unless footnote.attribute('id').value == id
-          description = []
-          paragraph_number = 0
-          footnote.xpath('w:p').each do |paragraph|
-            description << DocxParagraph.parse(paragraph, paragraph_number, default_paragraph, default_character)
-            paragraph_number += 1
-          end
-          return description
+    # @return [String] string for search of xpath
+    def xpath_for_search
+      "//w:#{path_suffix}"
+    end
+
+    # @return [String] string with xml path
+    def xml_path
+      "word/#{path_suffix}s.xml"
+    end
+
+    # Parse type and path suffix by node type
+    # @param [Nokogiri::XML:Node] node with HeaderFooter
+    def parse_type(node)
+      case node.name
+      when 'footnoteReference'
+        @path_suffix = 'footnote'
+        @type = :header
+      when 'endnoteReference'
+        @path_suffix = 'endnote'
+        @type = :footer
+      else
+        warn "Unknown HeaderFooter type: #{node.name}"
+      end
+    end
+
+    # Parse HeaderFooter
+    # @param [Nokogiri::XML:Node] node with HeaderFooter
+    # @return [HeaderFooter] result of parsing
+    def self.parse(node)
+      header_footer = HeaderFooter.new
+      header_footer.id = node.attribute('id').value.to_i
+      header_footer.parse_type(node)
+      doc = Nokogiri::XML(File.open(OOXMLDocumentObject.path_to_folder + header_footer.xml_path))
+      doc.search(header_footer.xpath_for_search).each do |footnote|
+        next unless footnote.attribute('id').value.to_i == header_footer.id
+        paragraph_number = 0
+        footnote.xpath('w:p').each do |paragraph|
+          header_footer.elements << DocxParagraph.parse(paragraph, paragraph_number, DocumentStructure.default_paragraph_style, DocumentStructure.default_run_style)
+          paragraph_number += 1
         end
       end
+      header_footer
     end
   end
 end